edge相关指令å分离
