+++ src/main/java/cn/edu/sjtu/ist/ecssbackendedge/controller/CommandController.java
@@ -1,7 +1,5 @@
 package cn.edu.sjtu.ist.ecssbackendedge.controller;
 
-import cn.edu.sjtu.ist.ecssbackendedge.model.scheduler.CollectScheduler;
-import cn.edu.sjtu.ist.ecssbackendedge.service.EdgeService;
 import cn.edu.sjtu.ist.ecssbackendedge.service.SensorService;
 import cn.edu.sjtu.ist.ecssbackendedge.utils.response.Result;
 import cn.edu.sjtu.ist.ecssbackendedge.utils.response.ResultUtil;
@@ -21,9 +19,6 @@ public class CommandController {
 
     @Autowired
     private SensorService sensorService;
-
-    @Autowired
-    private EdgeService edgeService;
 
     /**
      * 开启某个设备的sensor
@@ -45,35 +40,6 @@ public class CommandController {
     @PostMapping(value = "/sensor/stop/{id}/{sensorId}")
     public Result<?> stopSensor(@PathVariable String id, @PathVariable String sensorId) {
         return ResultUtil.success(sensorService.stopSensor(id, sensorId));
-    }
-
-    /**
-     * 开启边缘端数据收集的功能
-     * @param scheduler 调度器，间隔
-     * @return 成功/失败
-     */
-    @PostMapping("/edge/start")
-    public Result<?> startEdge(@RequestBody CollectScheduler scheduler) {
-        return ResultUtil.success(edgeService.startEdge(scheduler));
-    }
-
-    /**
-     * 关闭边缘端数据收集的功能
-     * @return 成功/失败
-     */
-    @PostMapping("/edge/stop")
-    public Result<?> stopEdge() {
-        return ResultUtil.success(edgeService.stopEdge());
-    }
-
-    /**
-     * 重启边缘端数据收集的功能
-     * @param scheduler 调度器，间隔
-     * @return 成功/失败
-     */
-    @PostMapping("/edge/restart")
-    public Result<?> restartEdge(@RequestBody CollectScheduler scheduler) {
-        return ResultUtil.success(edgeService.restartEdge(scheduler));
     }
 
     @PostMapping(value = "/start_monitor/{id}/{sensorId}")
